# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Şehit Hakan Gülşen Anadolu İHL Web Sitesi API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /news:
    get:
      operationId: getNews
      tags: [news]
      summary: Get all news
      responses:
        "200":
          description: News list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsListResponse"
    post:
      operationId: createNews
      tags: [news]
      summary: Create news (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNewsRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/News"
  /news/{id}:
    get:
      operationId: getNewsItem
      tags: [news]
      summary: Get news by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: News item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/News"
    put:
      operationId: updateNews
      tags: [news]
      summary: Update news (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNewsRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/News"
    delete:
      operationId: deleteNews
      tags: [news]
      summary: Delete news (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /students:
    get:
      operationId: getStudents
      tags: [students]
      summary: Get student posts
      responses:
        "200":
          description: Student posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentPostListResponse"
    post:
      operationId: createStudentPost
      tags: [students]
      summary: Create student post (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentPostRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentPost"
  /students/{id}:
    put:
      operationId: updateStudentPost
      tags: [students]
      summary: Update student post (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentPostRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentPost"
    delete:
      operationId: deleteStudentPost
      tags: [students]
      summary: Delete student post (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /teachers:
    get:
      operationId: getTeachers
      tags: [teachers]
      summary: Get all teachers
      responses:
        "200":
          description: Teachers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeacherListResponse"
    post:
      operationId: createTeacher
      tags: [teachers]
      summary: Create teacher (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeacherRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
  /teachers/{id}:
    put:
      operationId: updateTeacher
      tags: [teachers]
      summary: Update teacher (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeacherRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
    delete:
      operationId: deleteTeacher
      tags: [teachers]
      summary: Delete teacher (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    News:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        summary:
          type: string
        content:
          type: string
        imageUrl:
          type: string
        category:
          type: string
        publishedAt:
          type: string
        createdAt:
          type: string
      required: [id, title, summary, content, category, publishedAt, createdAt]
    NewsListResponse:
      type: object
      properties:
        news:
          type: array
          items:
            $ref: "#/components/schemas/News"
      required: [news]
    CreateNewsRequest:
      type: object
      properties:
        title:
          type: string
        summary:
          type: string
        content:
          type: string
        imageUrl:
          type: string
        category:
          type: string
        publishedAt:
          type: string
      required: [title, summary, content, category]
    StudentPost:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        authorName:
          type: string
        authorClass:
          type: string
        type:
          type: string
        imageUrl:
          type: string
        isStudentOfMonth:
          type: boolean
        createdAt:
          type: string
      required: [id, title, content, authorName, authorClass, type, isStudentOfMonth, createdAt]
    StudentPostListResponse:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/StudentPost"
      required: [posts]
    CreateStudentPostRequest:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        authorName:
          type: string
        authorClass:
          type: string
        type:
          type: string
        imageUrl:
          type: string
        isStudentOfMonth:
          type: boolean
      required: [title, content, authorName, authorClass, type]
    Teacher:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        branch:
          type: string
        title:
          type: string
        bio:
          type: string
        imageUrl:
          type: string
        videoUrl:
          type: string
        email:
          type: string
        createdAt:
          type: string
      required: [id, fullName, branch, title, createdAt]
    TeacherListResponse:
      type: object
      properties:
        teachers:
          type: array
          items:
            $ref: "#/components/schemas/Teacher"
      required: [teachers]
    CreateTeacherRequest:
      type: object
      properties:
        fullName:
          type: string
        branch:
          type: string
        title:
          type: string
        bio:
          type: string
        imageUrl:
          type: string
        videoUrl:
          type: string
        email:
          type: string
      required: [fullName, branch, title]
...